Updated code to work with OpenFOAM 6 and 8
There were several issues in running with both OpenFOAM 6 and 8 because
the 'stitchMesh' command creates a file in the latest time folder,
'<latestTime>/meshPhi' (or '0/meshPhi' if no time folders exist).
This messed up with the boundaries due to the several name changes
needed for repeatMesh.sh to work.

Solution was to remove all 'meshPhi' files after each 'stitchMesh'
operation.